Removed javascript proxy for write32 and read32 memory operations.
The old way broke with Emcripten. It would emit a LLVM trap when
a raw pointer in Rust was dereferenced in a specific way for a
memory write. So `write32` and `read32` were introduced which called
into pure JS code to do the 32-bit write and read since that provided
a work around; however, it consumed excessive CPU. Now, it has been
found that simply turning the write32 and read32 functions into pure
Rust code does not emit an LLVM trap and likely compiles away to an
inlined function (hopefully).